Create credentials
This is a known file that is used as an AWS configuration file that stores AWS secret access keys and it should
always be ignored and it should never be saved (with a commit) in the source version control system (Git in this
example). But let’s create this file for testing purpose.